requirements.txt
# ---------------------
# Deep Learning
# ---------------------
torch==2.0.1+cpu
torchvision==0.15.2+cpu
torchaudio==2.0.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

# ---------------------
# Web App
# ---------------------
streamlit>=1.29
numpy
pillow
opencv-python-headless
scikit-image
scikit-learn
matplotlib

# ---------------------
# ML Utils
# ---------------------
tqdm
pandas
albumentations
segmentation-models-pytorch==0.3.3

# ---------------------
# Descarga de modelos
# ---------------------
gdown
